requirements.txt
Django=5.2.3
psycopg2-binary=2.9.10
python-dotenv=1.1.0
langgraph=0.5.2
langsmith=0.4.5
langchain-groq=0.3.6
langchain-google-genai=2.1.7
langchain_mistralai=0.2.11
python-dotenv
ipython=9.4.0
colorama=0.4.6
pgvector=0.4.1
cohere=5.16.1
pinecone[asyncio]=7.3.0